improve the names of the two layouts for Ladin, making them clearer
Adding the specifiers "(Italian keyboard)" and "(German keyboard)"
may look over the top, as xkeyboard-config is all about keyboards,
but it makes things the most clear for the user.

Also, move the Ladin layout for the basic German keyboard to the file
for Italy, as the places where Ladin is spoken are all there.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>