symbols/ru: delete a redundant phonetic variant
The `phonetic_fr` variant differed from the `phonetic_azerty` variant
only by having some extra stuff on levels 3 + 4 (inhereted from `fr`).
Having extra stuff on the upper levels is not worth having a dedicated
variant for, especially not when this kind of layout is probably used
by just a handful of people.

So, just let the `phonetic_azerty` variant include the extra stuff
and drop the `phonetic_fr` variant, also because it's weird to have
a layout named "Russian (phonetic, French)" when one cannot type any
French on it.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>