symbols: replace most <LatK> key names with the stabler <AXnn> names
Key names of the form <AXnn> always refer to the same keys, whereas
names like <LatA>, <LatM>, and <LatZ> can refer to different keys,
depending on whether the first group is Qwerty, Azerty, or Qwertz.
For <LatA>, <LatQ>, <LatW>, <LatY>, and <LatZ> this is fine as they
get aliased always to other _letter_ keys.  But for <LatM> this is
troublesome as for Azerty layouts it gets aliased to <AC10>, which
in most layouts is a punctuation key, which means that either the
punctuation sign or the phonetic equivalent of the letter M is lost
when a "phonetic" layout is used together with the Azerty aliases.

Resolve this by using almost always the <AXnn> key names -- keeping
only the five non-troublesome aliases, where appropriate, in order to
not change anything for people who use a second-group phonetic layout
together with a first-group Qwertz layout.

This fixes issue #401.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>