symbols/in: rename a layout, after the language instead of the script
"Ol Chiki" is the name of the script [1] that is used to write the
Santali language [2].

  [1] https://en.wikipedia.org/wiki/Ol_Chiki_script
  [2] https://en.wikipedia.org/wiki/Santali_language

Naming the layout "Santali (Ol Chiki)" follows the examples of
"Punjabi (Gurmukhi)" and "Manipuri (Meitei)".  This is also how
the corresponding layouts are named in Android.

(Originally the comment near the beginning of `symbols/in` referred
to scripts, not languages, but commit b9971d9ba8 from three years ago
changed that.  Whether that was a good idea...)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>