symbols/my: harmonize the specification of the Zawgyi encoding
Zawgyi [1] is a font for the Burmese language where the characters
are encoded in a way that is incompatible with Unicode, so "Zawgyi"
is not part of the name of the language.  Also, the addition "Tai"
for a specific version of the font is unneeded.

  [1] https://en.wikipedia.org/wiki/Zawgyi_font

Furthermore, drop the `eng` marker for two layouts as no English
can be typed with them.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>