symbols/il: add the variant from the Israeli SI-1452-2 standard
Israeli standard SI-1452 part 2 (2018) describes a slight variation
of SI-1452 that removes almost all the deviations from the US layout.
But as it moves some symbols, it is added as an alternative layout.

For an image of the layout, see http://mikladot.com/.

(For conciseness, the variant defines only level 1 of the seven
changed keys, as nothing changes on the higher levels.)

This fulfills and closes issue #46.

Original-patch-by: Tzafrir Cohen <tzafrir@cohens.org.il>

Acked-by: Shai Berger <shai@platonix.com>

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>