set Arabic as the default layout for Egypt
Three years ago, commit 20d03d7441 added an exotic layout for Coptic,
and in the process created the `symbols/eg` file, putting Coptic as
the only and default and "basic" variant, although obviously it is
a niche layout (and was added to `rules/base.extras.xml`).

Arabic is the main and official language in Egypt.
So set the Arabic standard layout as the default.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>