tests: fix mktemp -d call on musl
mktemp -d can call mkdtemp(3), which:

The mkdtemp() function generates a uniquely named temporary directory
from template.  The last six characters of template must be XXXXXX and
these are replaced with a string that makes the directory name unique.

on glibc fewer characters are permitted, but musl hard requires the 6
X's as minimum.

avoids a:
  stderr:
  mktemp: mkdtemp failed on xkeyboard-config.XXXX: Invalid argument