symbols/it: change two acute accents to grave accents in Sicilian layout
Since around 2019 the Cademia Siciliana no longer uses acute accents
in the ortography of the Sicilian language, to make it easier to type
this language on the common Italian keyboard, which most speakers of
Sicilian use.

Thus drop the two key definitions with letters with acute accents,
so that they fall back to the grave-accented letters of the default
Italian layout.

Signed-off-by: David Paleino <david@cademiasiciliana.org>