symbols/br: remove the default Scroll_Lock mapping
This was added in 3b3d93cc06446a9555a1ef4649240c5f2d468308 but the
justification in the matching bugzilla entry [1] is:

  This way I wont need to put this in .bashrc:
     xmodmap -e 'add mod3 = Scroll_Lock'
  [...]
  Maybe it should be added earlier in other setup files like <pc>, but
  Im not sure. This way, more variants than <br> get it fixed, and the
  variants that needs to reconfigures mod3, will do it later.

IOW it looks like this was added to br(abnt2) for convenience despite
it not being part of the actual layout.

It was removed in !112 and later restored via !555 when that removal was
found to have the wrong reason.

The real reason is: there's nothing in the brazilian keyboard layout
that requires a scroll lock. Where it is useful to users it should be
enabled by those users via option "scrolllock:mod3".

[1] https://bugs.freedesktop.org/show_bug.cgi?id=28972