rules: generate the option->{symbols,compat,types} mappings from our XML files
The vast majority of options are a straightforward foo:bar = +foo(bar)
mapping. Instead of maintaining those mappings by hand let's generate
the rules file from the files we definitely maintain by hand: the
base.xml (and base.extras.xml) files.

This also makes it clearer which ones are exceptions and need to be
filled in through other means (or ignored because the option does not
affect any symbols).

The resulting rules file is identical to the previous hand-generated
one but it is alphabetically sorted and uses different whitespacing.

For merge.py to work we need to also fix the whitespaces in
rules/compat/0041-option_symbols.part, otherwise the duplicate header
detection won't work.