Improve description of caps:ctrl_modifier
The purpose of this option is to have Caps Lock behave as a Control
modifier, but still be able to distinguish it from the Control_L and
Control_R keys in things like games and emulators.

People tend to get confused when in some programs Caps Lock with this
option doesn't behave exactly like Control. ctrl:nocaps is probably the
option they really want.

Fixes #443