types: add the custom types file
This file "exists" thanks to a rules file entry to make it possible for
(non-libxkbcommon) users to install custom layouts with custom types.

Let's add it here to remove some special-casing we would otherwise need
in our generator scripts.

It must never be installed by us though as this would overwrite user
configuration.