rules: remove {lshift|rshift}_both_capslock{_cancel}
These were added as rules entries in ac8fa2656 but no-one noticed for 13
years that they had no corresponding base.xml entry. Let's remove them
because they look suspiciously like "added because we can" rather than
"added because there's a need for it". Google cannot find a single
reference to it that isn't this repo or a mirror of it.

Fixes #445