Add caps:digits_row option for Azerty layouts
The Azerty keyboards used in France and Belgium have the digits on the
shift level instead of on the base level of the top row. In Windows
the corresponding layouts have the curious characteristic that CapsLock
locks also the digits, meaning that some users have acquired the habit
of engaging CapsLock when they need to type some digits.

Fix this by adding a new option `caps:digits_row` that applies only to
the `$azerty` layouts group (currently `fr` and `be`). This option
changes the key type of the digits keys to `FOUR_LEVEL_SEMIALPHABETIC`.

There are two drawbacks with this solution:
- It is not possible to type the uppercase À, È, É and Ç. Without the
  option, one gets those by the internal transformation rules, done
  either by the X server or xkbcommon.
- It changes the behavior of the US variants from `fr`. But we do not
  expect anyone to use at the same time e.g. `fr` and `fr(us-alt)` with
  the new option `caps:number_row`. Instead we expect e.g. either one
  uses `fr(us-alt)` alone or `fr,us`. In the former case the option is
  not needed, while in the later case the option works correctly.