ru(ruu): "Ukrainian_i" as "Cyrillic_i" alternative
Add `Ukrainian_i` to the 3rd level of `<AB05>` (`Cyrillic_i`) for `ru(ruu)`.

RUU (Russian-Ukrainian layout) currently adds `Ukrainian_i` as an alternative
for `Cyrillic_yeru`. Arguably though, this letter is the Ukrainian/Belarusian
counterpart to the Russian "и", i.e. `Cyrillic_i`. There was no alternative for
it anyway, so this patch adds one _in addition to_ the one for `Cyrillic_yeru`.