Add missing characters from cp1253 and varEpsilon to Greek layout.
Most characters added are either present in the Windows layout, or
present in ISO-8859-7 (namely the ‘curly quotes’ added to level4 of
9 and 0), with the exception of Lunate Epsilon (`U03F5`) which was
added because it is an additonal graphical variant of the existing
Greek letter (like other additions already on the layout) and because
AltGr+Shift+E was freed up.
All characters have been attempted to be added to their respective
key positions from the Windows layout, or on level4 if level3 is
occupied.
Additionally, the Colemak variant of the layout is synced up with
the QWERTY variant.