Added Ergo‑L layout and variant
Ergo‑L is an ergonomic keyboard layout meant to optimize the typing
experience in French, English, and programming. Accents, diacritics
and fancy typographic symbols are typed using a custom dead key, but
it is fully implemented in the keymap itself as an `ISO_Level5_Latch.`

Below is a more detailed presentation of the layout (in French):

https://ergol.org/presentation/

This commit adds Ergo‑L `fr(ergol)`, along with an angle-mod variant
`fr(ergol_iso)` made for ISO keyboards.