rules: Fix layout compat rules
Layout compatibility rules are broken in configurations with 2 or more
layouts, e.g.:

- Works:
  - config: `--layout fi --variant basic`
  - symbols: `pc+fi(classic)+inet(evdev)`.
    `fi(basic)` is correctly remapped to `fi(classic)`.
- Does not work:
  - config `--layout fi,us --variant basic,`
  - symbols: `pc+fi(basic)+us:2+inet(evdev)`.
    `fi(basic)` is not remapped and leads to an include error.
- Does not work:
  - config `--layout us,fi --variant ,basic`
  - symbols: `pc+us:fi(basic):2+fi(classic):2+inet(evdev)`.
    `fi(basic)` is not remapped and leads to an include error.

I initially thought merging the following rules sets to fix it:

1. `model layout[n] = symbols`
2. `model layout[n] variant[n] = symbols`

but it does not work: `*` wildcard does *not* match empty values.
There is another issue: using a variant will check and match *both*
rules sets. In the first rule set, compat rules are only for deleted
symbols files (e.g. `dev`). But when the symbols file exists, we end
up with an invalid include statement (see `fi(basic)` examples above).
This is because the first rules set has a catch-all rule that will
always match, while we want it to match only in the second rules set.

Unfortunately such rules cannot be expressed in rules files. So the
solution is to append generated compatibility sections in the
corresponding symbols files and remove the corresponding current rules.

Section name clashes are checked and will raise an error.
E.g.: currently we append the following to `symbols/de`:

    partial xkb_symbols "lld" {
        include "it(lldde)"
    };

If we already had a section named lld, this would produce the following
error:

    Cannot add compatibility section in symbols/de: "lld" already exists

Also:
- Move symbols handling in meson into the symbols subfolder.
- Sort compat layout rules.
- We ensure that rules with variant take priority over rules without one.

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/728>