Fix NEC Japanese layout having two groups
According to our doc and practices, symbols should be defined only for
the first group. There are still a few layouts defining simultaneously
two groups. One of them is `nec_vndr/jp`, which has also a dedicated
rule. This commit fixes only this layout in preparation of a refactor
for `inet` rules.

- Split `nec_vndr/jp(pc98)` into two sections: `pc98` and `jp`, and
  make the latter the default section.
- Fix the rules to ensure using this layout with the model `pc98` will
  load *explicitly* 2 layouts, as this is the case for other models.
- Make the rules match `jp` layout, not `nec_vndr/jp`. No other rule
  used the vendor path to match a layout.

The remaining layouts with 2 groups should also be fixed someday.