triage: Add stale issues and MRs handling
Previous commits attempted to use bugbot to handle stale issues and MRs,
but bugbot only reacts on labels presence, it does not add them based on
some conditions.

The workflow is:
- An issue or MR is tagged automatically as stale after some period of
  inactivity, unless it has the label “skip: stale triage”.
- An issue or MR can be tagged manually as stale if it seems relevant
  to the maintainers. This is useful because the inactivity period in
  the previous section is quite long.
- A tagged stale issue or MR is closed automatically.
- When an issue or MR closed by the previous rules is re-opened, its
  labels related to its stale status are removed.

This commit adds a pipeline dedicated to triage, that is run only when
scheduled. The previous jobs are modified to run only if the pipeline is
*not* scheduled.