Add script to extract list of (non-)Latin layouts
This produces a CSV file with (non-)Latin layouts by analyzing all the
keysyms of each layout and check for required ones. This is more reliable
than checking the language tags of layouts, because a language may have
multiple scripts and the language tags may be incorrect or incomplete.

Method:
1. Get layouts via `xkbregistry`.
2. For each layout, check if it has all the basic Latin letters.
3. Export the filtered layouts as CSV.

Note that some layouts are “almost” Latin as they miss only a few
letters. Such layouts are considered non-Latin but raise a warning.