Update German keyboard according to DIN standard
In the August 2023 revision of the DIN 2137-1 standard,
some changes to the keyboard layers were made.

Most importantly, the third layer of the w, t, and i
keys was switched around.  That was already accounted for
in a previous commit.

The other changes were all in "group 2" of the keyboard.
Those changes are included in this commit.

A summary of the changes in the standard is freely available at
<https://www.dinmedia.de/de/norm/din-2137-1/353663175>, in German,
under "Änderungsvermerk".  Relevant for this commit are points f),
g), h), and i).