symbols/de: Add German (Noted) layout
Add the newest layout of the German Neo keyboard layout family. Noted is
a layout designed to make typing as comfortable and efficient as
possible in both German and English. It was developed with the help of a
layout optimizer, which is based on the metrics and concepts of
ArneBab's optimizer and adapts and extends them.

I switched from Neo 2 via Mine to Noted and I like to see this variant
to be available out of the box in Debian/Ubuntu.

Sources:
 * https://neo-layout.org/Layouts/noted/
 * https://dariogoetz.github.io/noted-layout/