symbols/inet: mark the F20-F23 mappings as historically special
These were used for mic mute and touchpad toggle keys because their
respective evdev keycodes end up being > 255. Recent evdev and libinput
drivers remap the evdev keycodes to F20-F23 as well so the rest of the
system doesn't have to.

Add a comment so we're not tempted to remove these mappings.

See also:
- https://github.com/systemd/systemd/pull/34325/
- https://gitlab.freedesktop.org/xorg/driver/xf86-input-libinput/-/merge_requests/62
- https://gitlab.freedesktop.org/xorg/driver/xf86-input-evdev/-/merge_requests/10