symbols/pc: Make "Alt and Meta are on Alt" disabled by default
The "Alt and Meta are on Alt" option is enabled by default very likely
due to the historical reasons. There used to be keyboards that had no
Windows or Super keys in the 90s, for example IBM Model M [1], so
providing some fallback seems reasonable. It seems there used to be
different ways to handle the lack of Super keys, e.g. using Escape or Alt
key.

According to [2], the Windows key made its first appearance in year 1994
and, with Windows 95, any keyboard with the "Designed for Windows" logo
had to have this key, which greatly contributed to making the Super key
widespread, it's hard to find a keyboard that has no Windows or Super key
nowadays.

From the desktop environment point of view, for the most optimal
experience, the user needs to have a keyboard with a real Super key (or a
spare key remapped to Super) because the desktop environments tend to
register global shortcuts containing Super key (with or without Shift
key).

The `altwin:meta_alt` option has been enabled for more than 20 years, but
it seems like users have started hitting its effects directly after
Plasma introduced the ability to configure global shortcuts that contain
only modifiers. Based on the empirical data that we (KDE developers) have
collected through various user forums and our bug tracker [3], the
current behavior is unexpected to our users, and it can interfere with
their workflow.

It is also interesting that the `altwin:meta_alt` option is effectively
enabled by default but there is no way to disable it. Perhaps there was
some mixup in 9bbda6bfbcc29a575ea3b5cb7bd2553767e80d53?

Last but not the least, the user expectations evolve as the time passes.
Today, practically all keyboards come with the Super key so users don't
face the problems that one would face 30 or 40 years ago (perhaps except
people who collect antique keyboards? but it is a niche case), so it is
reasonable to adjust the defaults to match the new expectations. The old
behavior can be restored by enabling the `altwin:meta_alt` option.

- [1] https://en.wikipedia.org/wiki/Super_key_(keyboard_button)
- [2] https://en.wikipedia.org/wiki/Windows_key
- [3] https://bugs.kde.org/show_bug.cgi?id=494579