Fix vendor rules
Some vendor rules match only on the layout but use the variant in
their symbols mapping. Such rules will make the keymap compilation
fail when a variant is not in the vendor-specific symbol file.

Example: the following rule

   ! model      layout            = symbols
     $applealu	$macvendorlayouts = macintosh_vndr/apple(alukbd)+macintosh_vndr/%l%(v)

used with the RMLVO (evdev; applealu_iso; gb; mac_intl) would produce
the following erroneous symbols include:

   macintosh_vndr/apple(alukbd)+macintosh_vndr/gb(mac_intl)

This fails because `mac_intl` is not a variant of `macintosh_vndr/gb`,
but of `gb`.

This commit fixes the issue by filtering vendor-specific variants,
having the vendor-agnostic symbols as a fallback.

Note that while it works perfectly for layouts in the first position,
layouts in later positions that have 1) a corresponding vendor-specific
file but 2) require a vendor-agnostic variant, will load the vendor-
specific symbols before importing the relevant section. To illustrate,
the RMLVO config (evdev; applealu_iso; ara,gb; ,mac_intl) would result
in the symbols:

    macintosh_vndr/apple(alukbd)+ara+macintosh_vndr/gb:2+gb(mac_intl):2

where `macintosh_vndr/gb:2` will be overriden by the following `gb(mac_intl):2`.
This is unfortunate, but there does not seem to be a workaround for it.

---

We would need to extend the rules syntax with a new wild card (e.g. ‘-’
or ‘∅’) to be able to match a missing variant, e.g.:

    ! model       layout[2]   variant[2] = symbols
      $applealu   gb          -          = +macintosh_vndr/gb:2
      $applealu   gb          *          = +%l[2]%(v[2]):2

While at it, another wild card to match both present *and* missing
variant would be really useful as well.

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/764>