Fix af layout with olpc model
Add `olpc` variant to `af` in order to fix compilation with the `olpc`
model. There are already various olpc vendor-specific variants in `af`
symbols. As the default variant for `af` is for the Dari language, make
also `olpc` variant an alias for the olpc Dari variant `fa-olpc`.

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/764>