Add hyper:mod4 option
Currently when using any option enabling `Hyper`, the virtual modifier
`Hyper` is mapped to `Mod3`. But this mapping conflicts with `LevelFive`
and leave layouts with 5+ levels unable to use both modifiers independently.

Added the option `hyper:mod4` to allow to choose the alternative mapping
of `Hyper` to `Mod4`. Note that this mapping conflicts with `Super`.

The rationale is:
- We have only a limited number of *real* modifiers available to use for
  the *virtual* modifiers[^1].
- `Hyper` has a niche use (mainly Emacs), so it should have lower priority
  than the use cases for `Super` and `LevelFive` (see hereinafter).
- `Super` is commonly used to control desktop environments, so it should
  always be functional by default.
- `LevelFive` is used in layouts with 5+ levels, so it should always be
  functional by default. However, it seems its use cases are less common
  than `Super`, so it seems safer to default `Hyper` to `Mod3/LevelFive`
  than to `Mod4/Super`.

[^1]: Note that this limit will be removed soon in libxkbcommon, allowing
      arbitrary independent modifiers.

Signed-off-by: Pierre Le Marre <dev@wismill.eu>
Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/760>