Use explicit virtual modifiers declarations
A virtual modifier should always be declared before its use. They are a
few missing entries, that currently work only because they are not used
in isolation.

But since XKB is about composability, we should ensure that every
component compiles in isolation, so this commit adds the missing
`virtual_modifiers` entries.

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/796>