Added KEY_ZENKAKUHANKAKU mapping for touchpad toggle.
Pressing the touchpad toggle key on some notebooks produces the key sequence
Super + Control + KEY_ZENKAKUHANKAKU¹. Actual Hankaku/Zenkaku keys however do
not use this scancode/keycode, but the tilde ones instead.

This patch binds the, until now on evdev unbound, KEY_ZENKAKUHANKAKU to touchpad
toggle so the desktop environments can act accordingly.

There is a small possibility of this being a breaking change: KDE on Wayland can
and does handle the Control + Super + KEY_ZENKAKUHANKAKU key combination
already. On X11 it is broken. In the next version it will also handle the
Control + Super + TouchpadToggle key combination, but, if xkeyboard-config with
this change is combined with an old KDE version without the corresponding
change, touchpad toggle will not work out of the box and the new key combination
needs to be added manually. With both updated however, it works on both X11 and
Wayland out of the box.

¹ According to this https://download.microsoft.com/download/1/6/1/161ba512-40e2-4cc9-843a-923143f3456c/translate.pdf
Hankaku/Zenkaku and F24 share a scancode. So it is quite possible that this is
a missmap by the linux kernel and the intent of the keyboards firmware devs was
to send Super + Control + F24.

Signed-off-by: Werner Sembach <wse@tuxedocomputers.com>
Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/810>