meson: Use versioning scheme
Before this commit our installed files were not versioned,
so only one version of xkeyboard-config could be installed at a time.

We plan to introduce breaking changes to the rules and keymap file
formats to fix long-standing issues. To ensure backward compatibility,
we are switching to versioned install directories and files, in order
to enable multiple versions of xkeyboard-config to be installed in
parallel.

Since we do not expect the formats to evolve frequently, the version
scheme is simple:
- The file format version uses a single number and increments by 1.
- The *major* version of the xkeyboard-config project denotes the
  file format version of *both* rules and keymap files.
- The version scheme starts at the current major version: **2**.
- The version defines a range of XKB compilers that it is compatible
  with.
- Everytime we introduce a breaking change in the file formats, we bump
  the major version of xkeyboard-config and define the corresponding
  supported XKB compilers range.

This commit only introduces the version scheme, in order to give time
to distributions and downstream to adapt.

Foreseen changes:
- v3: Introduction of the new wild cards in the rules.
- v4: Introduction of new parameters for lock actions, in order to
  fix the infamous XKB limitation “need to kick hotkeys on release, not
  press”.

TODO:
- doc: write more formal policies & distribution guidelines

Part-of: <https://gitlab.freedesktop.org/xkeyboard-config/xkeyboard-config/-/merge_requests/817>